+++ src/index.ts
@@ -1,6 +1,6 @@
-export { Column, Entity, Table, getEntities, getTableName } from "./decorators";
+export { Column, Entity, Table, getEntities, getTables, getTableName } from "./decorators";
 export { ITypedQueryBuilder, TypedKnex, registerBeforeInsertTransform, registerBeforeUpdateTransform } from "./typedKnex";
-export { validateEntities, validateTables } from "./validateEntities";
+export { validateEntities, validateTables } from "./validateTables";
 export { FlattenOption } from "./unflatten";
 export { ICustomDatabaseType } from "./ICustomDatabaseType";
 export { registerQueryBuilderExtension } from "./registerQueryBuilderExtension";