+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+-   Chore: Deprecate `primary` property of column.
 -   Feature: Added `getSingleOrUndefined` and `getFirstOrUndefined`.
 
 ## 4.8.4