+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+-   Chore: Properly deprecate mentions of "entity".
 -   Chore: Deprecate `primary` property of column.
 -   Feature: Added `getSingleOrUndefined` and `getFirstOrUndefined`.