+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 4.9.0
+
 -   Chore: Use prepack script instead of prepublish.
 -   Chore: Properly deprecate mentions of "entity".
 -   Chore: Deprecate `primary` property of column.